:up: replace actions in column with action props
